Redefine types
The types now better match the tables in the db. Having Routine exercises contain the exercise object removes the need of an api call to ninjas api in the backend.
