+++ src/shared/types/Exercise.ts
@@ -1,0 +1,20 @@
+export type Exercise = {
+  name: string,
+  type: string,
+  muscle: string,
+  equipment: string,
+  difficulty: string,
+  instructions: string,
+};
+
+const initExercise  = {
+  name: '',
+  type: '',
+  muscle: '',
+  equipment: '',
+  difficulty: '',
+  instructions: ''
+};
+
+export { initExercise };
+
