+++ src/shared/types/Login.ts
@@ -1,0 +1,4 @@
+export type Login = {
+  email: string,
+  password: string,
+};
