+++ src/shared/types/Routine.ts
@@ -1,0 +1,6 @@
+import { RoutineExercise } from './RoutineExercise';
+
+export type Routine = {
+  name: string,
+  exercises: RoutineExercise[]
+};
