+++ src/shared/types/RoutineExercise.ts
@@ -1,0 +1,9 @@
+import { Exercise } from './Exercise';
+
+export type RoutineExercise = {
+  exercise: Exercise
+  weight: string;
+  sets: number;
+  quantity: number;
+  quantityUnit: string;
+}
