+++ src/shared/types/Signup.ts
@@ -1,0 +1,5 @@
+export type Signup = {
+  name: string,
+  email: string,
+  password: string,
+};
